+++ tests/CompilerTest.php
@@ -51,22 +51,22 @@
                                * You should not modify it.
                                *
                                * @i18n-package foo
-                               * @i18n-version 1.0.0
+                               * @i18n-version 2.0.0
                                * @createdAt 2021-08-22T13:37:42+00:00
                                */
                               
                               return [
                                   'meta' => [
-                                      'i18n-version' => [1, 0, 0]
+                                      'i18n-version' => [2, 0, 0]
                                   ],
                                   'keys' => [
-                                      'bar' => [fn(array $context) => "I am a test key", 'I am a test key'],
+                                      'bar' => 'I am a test key',
                                       'hello.world' => [fn(array $context) => "Hello " . ($context['name'] ?? '') . "!", 'Hello {name}!'],
                                       'with.linefeed' => [fn(array $context) => "Hello\n" . ($context['name'] ?? '') . "!", 'Hello'."\n".'{name}!'],
                                       'with.doubleQuote' => [fn(array $context) => "Hello \"" . ($context['name'] ?? '') . "\"!", 'Hello "{name}"!'],
                                       'with.singleQuoteInVariable' => [fn(array $context) => "Hello " . ($context['name\'lol'] ?? '') . "!", 'Hello {name\'lol}!'],
                                       'with.linefeedInVariable' => [fn(array $context) => "Hello " . ($context['name\'."\n".\'lol'] ?? '') . "!", 'Hello {name'."\n".'lol}!'],
-                                      'with.variableToken' => [fn(array $context) => "Hello \$name!", 'Hello $name!'],
+                                      'with.variableToken' => 'Hello $name!',
                                   ]
                               ];